utl: move generic RAII mechanism to utl
